Business model approach removed.
Defining the operational model of LabSpace
as a business model wasn't a good approach.
We tried to define things that in different
contexts suffer from great variability so the
new approach involves studies performed on the
different stakeholders and a planning phase
to design LabSpace specificaly for the context
of implementation.